breaking: consolidated supabase_state, and supabase_auth_required_state to supabase_auth_state
